requirements.txt
auto-gptq
auto_gptq-0.4.1+cu118-cp310-cp310-linux_x86_64.whl
beautifulsoup4==4.12.2
chromedriver-autoinstaller==0.6.2
ctransformers==0.2.27
faiss-cpu==1.7.4
Flask
InstructorEmbedding==1.0.1
jinja2==3.0.3
langchain==0.0.251
llama-cpp-python==0.2.6
llama-index=0.8.36
mlflow==2.7.1
nemoguardrails==0.5.0
openai==0.27.8
pandas==2.1.1
pdfminer-six==20221105
pyvis==0.3.2
PyMuPDF==1.23.3
pysqlite-binary
requests==2.31.0
selenium==4.12.0
tiktoken==0.4.0
transformers==4.33.2
watchdog==3.0.0
#upgrade setuptools
sentence-transformers
streamlit==1.25.0
torch==2.0.1
xformers==0.0.20
